Avoid redundant indexing.
Previously we would re-index the same jars for every project/config and
on every configuration change. Now we only index each jar once.
