+++ metaserver/src/main/scala/scala/meta/languageserver/Jars.scala
@@ -5,21 +5,27 @@ import java.io.PrintStream
 import coursier._
 import org.langmeta.io.AbsolutePath
 
+case class ModuleID(organization: String, name: String, version: String) {
+  def toCoursier: Dependency = Dependency(Module(organization, name), version)
+  override def toString: String = s"$organization:$name:$version"
+}
 object Jars {
-  // TODO(olafur) this method should return a monix.Task.
   def fetch(
       org: String,
       artifact: String,
       version: String,
       out: PrintStream,
       sources: Boolean = false
+  ): List[AbsolutePath] =
+    fetch(ModuleID(org, artifact, version) :: Nil, out, sources)
+
+  def fetch(
+      modules: Iterable[ModuleID],
+      out: PrintStream,
+      sources: Boolean
   ): List[AbsolutePath] = {
     val classifier = if (sources) "sources" :: Nil else Nil
-    val res = Resolution(
-      Set(
-        Dependency(Module(org, artifact), version)
-      )
-    )
+    val res = Resolution(modules.toIterator.map(_.toCoursier).toSet)
     val repositories = Seq(
       Cache.ivy2Local,
       MavenRepository("https://repo1.maven.org/maven2")
@@ -57,5 +63,4 @@ object Jars {
       jars.map(AbsolutePath(_))
     }
   }
-
 }
