+++ test-workspace/build.sbt
@@ -1,8 +1,8 @@
 inThisBuild(
   List(
-    scalaVersion := "2.12.3",
+    scalaVersion := "2.12.4",
     addCompilerPlugin(
-      "org.scalameta" % "semanticdb-scalac" % "2.0.1" cross CrossVersion.full
+      "org.scalameta" % "semanticdb-scalac" % "2.1.1" cross CrossVersion.full
     ),
     libraryDependencies +=
       ("org.scalatest" %% "scalatest" % "3.0.3" % Test).withSources(),
